Rename nompi -> serial
